requirements.txt
blinker @ file:///C:/b/abs_b1i87khtob/croot/blinker_1737448732095/work
certifi==2025.4.26
charset-normalizer==3.4.2
click @ file:///C:/b/abs_42rbjngfys/croot/click_1744271606983/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
Flask @ file:///C:/b/abs_a6m37uz578/croot/flask_1737454335662/work
fonttools==4.58.0
fsspec==2025.5.1
gunicorn==23.0.0
idna==3.10
itsdangerous @ file:///C:/b/abs_c4vwgdr5yn/croot/itsdangerous_1716533399914/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow @ file:///C:/b/abs_68t826txdy/croot/pillow_1744613085333/work
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.3
six==1.17.0
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.145
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
